requirements.txt
Django==4.2
djangorestframework==3.14
django-cors-headers==4.3
Pillow==10.2
firebase-admin==6.4
python-dateutil==2.8
cloudinary==1.36
django-cloudinary-storage==0.3.0
gunicorn==21.2
whitenoise==6.6
dj-database-url==2.1
psycopg2-binary==2.9.9
